% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenererArbresBeta.R
\name{GenererArbresBeta}
\alias{GenererArbresBeta}
\title{Fonction permettant de générer des arbres.}
\usage{
GenererArbresBeta(shp, gha, seuilDist, beta1, beta2, dmax, dmin)
}
\arguments{
\item{shp}{= Périmètre sous forme d'un objet sf}

\item{gha}{= Surfacec terrière à l'hectare}

\item{beta1}{= premier paramètre d'un loi béta}

\item{beta2}{= second paramètre d'un loi béta}

\item{dmax}{= diam max}

\item{dmin}{= diam min}

\item{SeuilDist}{= distance mini acceptable entre tiges}
}
\value{
Une liste d'arbres sous forme d'un objet sf
}
\description{
Fonction permettant de générer des arbres à partir d'une surface terrière objectif
et d'une distribution des diamètres selon une loi Béta définie par les paramètres beta1 et beta2.
}
\examples{
library(sf, quietly =T, warn.conflicts =F)
library(tidyverse, quietly =T, warn.conflicts =F)
library(units)
library(Generer)
# ----------------- Données
SeuilDist  <- set_units(1, m)    # Seuil minimum de distance entre 2 tiges
Beta1      <- 1    # coefft loi beta
Beta2      <- 4    # coefft loi beta
Gha        <- 17   # surface terriere objectif (m2/ha)
Dmax       <- 80
Dmin       <- 18
# ----------------- creation arbres
data(perim)
surf <- as.vector(st_area(perim))/10000
arbres <- GenererArbresGD(perim, Gha, SeuilDist, Beta1, Beta2, Dmax, Dmin)
sum(arbres$G)/surf

}
\author{
Bruciamacchie Max
}
