% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenererPlaCirc.R
\name{GenererPlaCirc}
\alias{GenererPlaCirc}
\title{Fonction permettant de générer des placettes circulaires}
\usage{
GenererPlaCirc(shp, arbres, rayon = 15, NbPlac = 10, type = "regular")
}
\arguments{
\item{shp}{= Périmètre sous forme d'un objet sf}

\item{arbres}{= fichier obtenu avec la fonction GenererArbres.}

\item{rayon}{= rayon de la placette en mètre. Par défaut rayon = 15 m.}

\item{NbPlac}{= nombre de placettes. Par défault NbPlac = 10.}

\item{type}{= type de répartition. Par défault type = "regular".
Le type peut également être "random" ou "hexagonal".
Dans le cas du type "random" le nombre exact de placettes sera tiré au hasard.
Pour les autres types, le nombre de placettes sera approché.}
}
\value{
Une liste da placettes sous forme d'un objet sf.
}
\description{
Fonction permettant de générer n placettes circulaires sur un périmètre. Leur répartition
peut être régulière ou aléatoire
}
\examples{
data(perim)
res <- Generer::GenererPlaCirc(perim, arbres)
res$Estim
res$graph

}
\author{
Bruciamacchie Max
}
