% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenererPlaPCQM.R
\name{GenererPlaPCQM}
\alias{GenererPlaPCQM}
\title{Fonction permettant de générer des placettes PCQM}
\usage{
GenererPlaPCQM(
  shp,
  placettes,
  arbres,
  rayonPCQM = 25,
  NbPlac = 10,
  type = "regular"
)
}
\arguments{
\item{shp}{= Périmètre sous forme d'un objet sf}

\item{arbres}{= fichier obtenu avec la fonction GenererArbres.}

\item{rayonPCQM}{= rayon maximal de recherche des arbres. Par défault rayonPCQM = 25m.}

\item{NbPlac}{= nombre de placettes. Par défault NbPlac = 10.}

\item{type}{= type de répartition. Par défault type = "regular".
Le type peut également être "random" ou "hexagonal".
Dans le cas du type "random" le nombre exact de placettes sera tiré au hasard.
Pour les autres types, le nombre de placettes sera approché.}
}
\value{
Une liste da placettes sous forme d'un objet sf.
}
\description{
Fonction permettant de générer n placettes PCQM.
Leur répartition peut être régulière ou aléatoire
}
\examples{
data(perim)
res <- Generer::GenererPlaPCQM(perim, arbres)
res$Estim
res$graph

}
\author{
Bruciamacchie Max
}
