% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculPCQM.R
\name{CalculPCQM}
\alias{CalculPCQM}
\title{Calcul placettes PCQM}
\usage{
CalculPCQM(tab, pla)
}
\arguments{
\item{tab}{= tableau}
}
\value{
Surface terrière et densité moyennes.
}
\description{
Calcul de la surface terrière et de la densité à partir n placettes PCQM.
}
\details{
Le tableau en entrée doit contenir 4 colonnes dénommées "NumPlac","Quart","rayon","Diam".
}
\examples{
data(perim)
res <- Generer::GenererPlaPCQM(perim, arbres)
res$Estim
res$graph

}
\author{
Bruciamacchie Max
}
